Update protobufjs to 6.8.8
Protobuf.js v6 is breaking change with alterations to the API
in several places.

Because we are contemplating a protocol rewrite, this drops the
dependency on 'pb-stream', which used the old API in favor of
inlining the encoder/decoder logic with changes.

This does not affect the message wire-format, and so does not
cause issues of client/server compatibility.